Split wasm source
