+++ taichi/runtime/wasm/aot_module_builder_impl.cpp
@@ -1,4 +1,4 @@
-#include "taichi/backends/wasm/aot_module_builder_impl.h"
+#include "taichi/runtime/wasm/aot_module_builder_impl.h"
 #include "taichi/llvm/llvm_program.h"
 
 #include <fstream>
